fix: move all from plotmixin to base document
