+++ docarray/base_document/mixins/__init__.py
@@ -1,4 +1,3 @@
-from docarray.base_document.mixins.plot import PlotMixin
 from docarray.base_document.mixins.proto import ProtoMixin
 
-__all__ = ['PlotMixin', 'ProtoMixin']
+__all__ = ['ProtoMixin']
